Resolved AccountManager Issue, gets request moved to repository
