+++ backend/Accounts/PetFemily.Accounts.Application/IAccountReadRepository.cs
@@ -1,0 +1,12 @@
+﻿using PetFemily.Accounts.Application.Dto;
+
+namespace PetFemily.Accounts.Application;
+
+public interface IAccountReadRepository
+{
+    Task<UserDto> GetFullInfoUserByIdAsync(
+        Guid userId,
+        CancellationToken cancellationToken);
+
+    Task<bool> UserExistByIdAsync(Guid userId, CancellationToken cancellationToken);
+}
